+++ Demo/src/main/java/com/braintreepayments/demo/PayPalFragment.java
@@ -17,7 +17,9 @@ import androidx.navigation.fragment.NavHostFragment;
 import com.braintreepayments.api.DataCollector;
 import com.braintreepayments.api.PayPalClient;
 import com.braintreepayments.api.PayPalLauncher;
+import com.braintreepayments.api.PayPalPaymentAuthRequest;
 import com.braintreepayments.api.PayPalRequest;
+import com.braintreepayments.api.PayPalResult;
 import com.braintreepayments.api.PaymentMethodNonce;
 
 public class PayPalFragment extends BaseFragment {
@@ -44,11 +46,11 @@ public class PayPalFragment extends BaseFragment {
         payPalClient = new PayPalClient(requireContext(), super.getAuthStringArg());
         payPalLauncher = new PayPalLauncher(
                 paymentAuthResult -> payPalClient.tokenize(
-                        paymentAuthResult, (payPalAccountNonce, error) -> {
-                            if (error != null) {
-                                handleError(error);
-                            } else if (payPalAccountNonce != null) {
-                                handlePayPalResult(payPalAccountNonce);
+                        paymentAuthResult, (payPalResult) -> {
+                            if (payPalResult instanceof PayPalResult.Failure) {
+                                handleError(((PayPalResult.Failure) payPalResult).getError());
+                            } else if (payPalResult instanceof PayPalResult.Success) {
+                                handlePayPalResult(((PayPalResult.Success) payPalResult).getNonce());
                             }
                         }));
 
@@ -97,11 +99,12 @@ public class PayPalFragment extends BaseFragment {
             payPalRequest = createPayPalCheckoutRequest(activity, amount);
         }
         payPalClient.createPaymentAuthRequest(activity, payPalRequest,
-                (paymentAuthRequest, error) -> {
-                    if (error != null) {
-                        handleError(error);
-                    } else {
-                        payPalLauncher.launch(requireActivity(), paymentAuthRequest);
+                (paymentAuthRequest) -> {
+                    if (paymentAuthRequest instanceof PayPalPaymentAuthRequest.Failure) {
+                        handleError(((PayPalPaymentAuthRequest.Failure) paymentAuthRequest).getError());
+                    } else if (paymentAuthRequest instanceof PayPalPaymentAuthRequest.ReadyToLaunch){
+                        payPalLauncher.launch(requireActivity(),
+                                ((PayPalPaymentAuthRequest.ReadyToLaunch) paymentAuthRequest).getRequestParams());
                     }
                 });
     }